% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rewlr.R
\name{predict.rewlr}
\alias{predict.rewlr}
\title{predicting the rare event weighted logistic regreression}
\usage{
\method{predict}{rewlr}(model, newdata = NULL)
}
\arguments{
\item{model}{an object of class "rewlr"}

\item{newdata}{a newdata to be predicted}
}
\value{
return the predicted by the model in probablity form.
}
\description{
preditct method for class "rewlr".
}
